+++ relay-general/src/store/mod.rs
@@ -22,7 +22,9 @@ mod trimming;
 
 pub use self::clock_drift::ClockDriftProcessor;
 pub use self::geo::{GeoIpError, GeoIpLookup};
-pub use normalize::breakdowns::BreakdownsConfig;
+pub use normalize::breakdowns::{
+    get_breakdown_measurements, BreakdownConfig, BreakdownsConfig, SpanOperationsConfig,
+};
 pub use normalize::normalize_dist;
 
 /// The config for store.